Updating SDL and SDL_Image to the latest SDL2 builds, plus updating libpng to the most recent version.

+++ classes/jedi-sdl/SDL/Pas/sdlstreams.pas
@@ -115,7 +115,7 @@ begin
 end;
 *)
 
-function SdlStreamSeek( context : PSDL_RWops; offset : Integer; whence : Integer ) : integer; cdecl;
+function SdlStreamSeek( context : PSDL_RWops; offset : Int64; whence : Integer ) : Int64; cdecl;
 var
   stream : TStream;
   origin : Word;
